+++ docs/changelog.rst
@@ -6,7 +6,7 @@ Changelog
 * :bug:`-` Transfers between tracked accounts will now have a correct label in the UI.
 * :bug:`-` Users will be able to finish balance queries if they have assets with missing information.
 * :bug:`5038` Premium users with big databases should no longer see the error: "Upload data to server died with exception: database plaintext is locked"
-
+* :bug:`-` If a user removes the API keys for an exchange, actions on that exchange will no longer be excluded from PnL reports.
 
 * :release:`1.26.2 <2022-12-09>`
 * :bug:`5187` Fix issue where the blockchain balances were refreshed, when only changing the label or the tags of an account.